requirements.txt
django==6.0.3
djangorestframework==3.17.1
drf-spectacular==0.29.0
